requirements.txt
boto3
pandas
pygsheets
python-dateutil
python-dotenv
PyYAML
splitwise
